Fixed sigsegv - mingw++ on Windows 32bit only
Closing a trace file caused a sigsegv when compiled with mingw++ 32bit on WIndows 7.
Caused by the deletion of map entries for binds for the cursor being closed.
Interesting one to track down though!

+++ TraceMiner2/tmoptions.cpp
@@ -78,7 +78,7 @@ bool tmOptions::parseArgs(int argc, char *argv[]) {
         // Convert args to lower case and to strings.
         // Not really safe with Unicode though!
         string thisArg = string(argv[arg]);
-        for (int c = 0; c < thisArg.length(); c++) {
+        for (int c = 0; c < (int)thisArg.length(); c++) {
             thisArg[c] = tolower(thisArg[c]);
         }
